LD (nnnn),A implemented, need peripherals
CPU now makes it as far as instruction 0064h in the boot rom without
obvious error. At this point, the code enters a loop and tests
the register at ff44, which depends on the status of the LCD screen.
Setting this value to 90h immediately drops out of the loop,
allowing further debugging.